add _form partial for create and edit form

+++ src/Controller/PinsController.php
@@ -71,10 +71,12 @@ class PinsController extends AbstractController
     }
 
     #[Route("/pins/{id<[0-9]+>}/delete", name:"app_pins_delete", methods:["DELETE"])]
-    public function delete(Pin $pin):Response
+    public function delete(Pin $pin, Request $request):Response
     {
-        $this->em->remove($pin);
-        $this->em->flush();
+        if($this->isCsrfTokenValid('pin_deletion_'.$pin->getId(), $request->request->get('csrf_token'))){
+            $this->em->remove($pin);
+            $this->em->flush();
+        }
         return $this->redirectToRoute("app_home");
     }
 }
